examples/fastapi-grpc-client/requirements.txt
fastapi
uvicorn[standard]
grpcio
grpcio-tools
